% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtlr-package.R
\docType{package}
\name{rtlr-package}
\alias{rtlr}
\alias{rtlr-package}
\title{rtlr: Print Right-to-Left Languages Correctly}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Convenience functions to make some common tasks with right-to-left string printing easier, more convenient and with no need to remember long Unicode characters. Specifically helpful for right-to-left languages such as Arabic, Persian and Hebrew.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matanhakim/rtlr}
  \item \url{https://matanhakim.github.io/rtlr/}
  \item Report bugs at \url{https://github.com/matanhakim/rtlr/issues}
}

}
\author{
\strong{Maintainer}: Matan Hakim \email{matanhakim@gmail.com} (\href{https://orcid.org/0009-0002-9372-8975}{ORCID}) [copyright holder]

}
\keyword{internal}
