% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_rtl.R
\name{str_rtl}
\alias{str_rtl}
\title{Wrap a string with RTL-embedding Unicode characters}
\usage{
str_rtl(..., multiline = FALSE)
}
\arguments{
\item{...}{One or more character vectors, ideally of length 1.}

\item{multiline}{Whether to split the input strings to separate lines. Defaults to \code{FALSE}.}
}
\value{
A character vector.
}
\description{
Wrap a string with RTL-embedding Unicode characters
}
\examples{
# Notice how the `.` prints to the right, while it should print to the left
cat("א.")

# `str_rtl()` fixes this:
str_rtl("א.") |> cat()

str_rtl("א.", "ב.", multiline = TRUE) |> cat()
}
